in sample schema changed individual from prop to calc prop
