+++ src/encoded/schemas/sample.json
@@ -22,6 +22,9 @@
         "accession" : {
             "accessionType" : "SA"
         },
+        "schema_version": {
+            "default": "2"
+        },
         "workup_type": {
             "title": "Workup Type",
             "type": "string",
@@ -30,12 +33,6 @@
                 "WES",
                 "Sanger"
             ]
-        },
-        "individual": {
-            "title": "Individual",
-            "description": "Individual this sample is derived from",
-            "type": "string",
-            "linkTo": "Individual"
         },
         "specimen_type": {
             "title": "Specimen Type",
